+++ flask_split/core.py
@@ -120,15 +120,20 @@ def finished(experiment_name, reset=True):
             return
         alternative_name = _get_session().get(experiment.key)
         if alternative_name:
-            if '%s:finished' % experiment.key not in _get_session():
+            if 'split_finished' not in session:
+                session['split_finished'] = set()
+            if experiment.key not in session['split_finished']:
                 alternative = Alternative(alternative_name, experiment_name)
                 alternative.increment_completion()
             if reset:
                 _get_session().pop(experiment.key, None)
-                _get_session().pop('%s:finished' % experiment.key, None)
+                try:
+                    session['split_finished'].remove(experiment.key)
+                except KeyError:
+                    pass
                 session.modified = True
             else:
-                _get_session()['%s:finished' % experiment.key] = 1
+                session['split_finished'].add(experiment.key)
                 session.modified = True
     except ConnectionError:
         if not current_app.config['SPLIT_DB_FAILOVER']: